correccion de Authmodule y jwt strategy
